# docs/swagger.yaml
basePath: /api/v2
definitions:
...
  internal_handlers.createWalletRequest:
    properties:
      address:
        description: optional — derived from private_key if omitted
        type: string
      name:
        type: string
      private_key:
        type: string
    type: object
...
  internal_handlers.walletResponse:
    properties:
      address:
        type: string
      created_at:
        type: string
      gas_balance:
        type: string
      id:
        type: integer
      is_default:
        type: boolean
      name:
        type: string
      payment_balance:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Enterprise gateway for Autonomi decentralized storage. Provides file
    upload/download, user management, API tokens, webhooks, and admin controls.
  title: Indelible API
  version: "2.0"
paths:
...
  /admin/wallets:
    get:
      description: Return all configured wallets with their balances
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/internal_handlers.walletResponse'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all wallets
      tags:
      - 'Admin: Wallets'
    post:
      consumes:
      - application/json
      description: Add a new wallet with encrypted private key storage
      parameters:
      - description: Wallet details including private key
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.createWalletRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a wallet
      tags:
      - 'Admin: Wallets'
  /admin/wallets/{id}:
    delete:
      description: Remove a wallet (cannot delete the default wallet)
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a wallet
      tags:
      - 'Admin: Wallets'
  /admin/wallets/{id}/balance:
    post:
      description: Query the EVM chain for the current token and gas balance
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Refresh wallet balance
      tags:
      - 'Admin: Wallets'
  /admin/wallets/{id}/default:
    put:
      description: Make a wallet the default for upload payments
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Set default wallet
      tags:
      - 'Admin: Wallets'
  /admin/wallets/{id}/transactions:
    get:
      description: List recorded transactions for a wallet (upload payments, refunds),
        newest first. V2-321.
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: integer
      - description: Max results (default 50, max 100)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Per-wallet transaction history
      tags:
      - 'Admin: Wallets'
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT or API token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"